- Add context menu item for documentDB, graph and table - DocDBAcctTreeItemBase: add delete acct for docdb, graph and table - commands.ts -> folder with one command per file, and its imports
