+++ src/extension.ts
@@ -27,7 +27,7 @@ import { DocDBAccountTreeItemBase } from './docdb/tree/DocDBAccountTreeItemBase'
 import { GraphAccountTreeItem } from './graph/tree/GraphAccountTreeItem';
 import { DocDBAccountTreeItem } from './docdb/tree/DocDBAccountTreeItem';
 import { TableAccountTreeItem } from './table/tree/TableAccountTreeItem';
-import { deleteCosmosDBAccount } from './commands';
+import { deleteCosmosDBAccount } from './commands/deleteCosmosDBAccount';
 
 export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(new Reporter(context));
@@ -58,7 +58,7 @@ export function activate(context: vscode.ExtensionContext) {
 	});
 	actionHandler.registerCommand('cosmosDB.deleteAccount', async (node?: IAzureNode) => {
 		if (node) {
-			await deleteCosmosDBAccount(node);
+			node.deleteNode();
 		}
 	});
